+++ src/components/PurchaseOrder/PO.js
@@ -11,6 +11,7 @@ import {
 } from '@folio/stripes/core';
 import {
   baseManifest,
+  LIMIT_MAX,
   Tags,
   TagsBadge,
   useModalToggle,
@@ -42,6 +43,7 @@ import {
   FUND,
   LINES_LIMIT,
   ORDER,
+  ORDER_INVOICES,
 } from '../Utils/resources';
 import {
   cloneOrder,
@@ -72,6 +74,7 @@ const PO = ({
   const [handleErrorResponse] = useHandleOrderUpdateError(mutator.expenseClass);
 
   const [order, setOrder] = useState();
+  const [orderInvoicesIds, setOrderInvoicesIds] = useState();
   const [isLoading, setIsLoading] = useState(true);
   const [isErrorsModalOpened, toggleErrorsModal] = useModalToggle();
   const [updateOrderErrors, setUpdateOrderErrors] = useState();
@@ -87,13 +90,26 @@ const PO = ({
 
   const fetchOrder = useCallback(
     () => mutator.orderDetails.GET()
-      .then(setOrder)
-      .catch(() => {
+      .then(orderResp => {
+        setOrder(orderResp);
+
+        return mutator.orderInvoicesRelns.GET({
+          params: {
+            query: `purchaseOrderId==${orderResp.id}`,
+            limit: LIMIT_MAX,
+          },
+        });
+      }, () => {
         sendCallout({
           message: <SafeHTMLMessage id="ui-orders.errors.orderNotLoaded" />,
           type: 'error',
         });
-      }),
+      })
+      .then(orderInvoicesResp => {
+        const invoicesIds = orderInvoicesResp.map(({ invoiceId }) => invoiceId);
+
+        setOrderInvoicesIds(invoicesIds);
+      }, () => setOrderInvoicesIds([])),
     // eslint-disable-next-line react-hooks/exhaustive-deps
     [match.params.id, sendCallout],
   );
@@ -510,7 +526,10 @@ const PO = ({
             id="POSummary"
             label={<FormattedMessage id="ui-orders.paneBlock.POSummary" />}
           >
-            <SummaryView order={order} />
+            <SummaryView
+              order={order}
+              orderInvoicesIds={orderInvoicesIds}
+            />
           </Accordion>
           <Accordion
             displayWhenOpen={addPOLineButton}
@@ -525,7 +544,7 @@ const PO = ({
           </Accordion>
           <POInvoicesContainer
             label={<FormattedMessage id="ui-orders.paneBlock.relatedInvoices" />}
-            orderId={match.params.id}
+            orderInvoicesIds={orderInvoicesIds}
           />
         </AccordionSet>
       </AccordionStatus>
@@ -614,6 +633,11 @@ PO.manifest = Object.freeze({
     ...baseManifest,
     accumulate: true,
     fetch: false,
+  },
+  orderInvoicesRelns: {
+    ...ORDER_INVOICES,
+    fetch: false,
+    accumulate: true,
   },
 });